requirements.txt
protobuf>=4.25.3,<5.0
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
comtypes==1.4.11
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
deepseek==1.0.0
defusedxml==0.7.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.0
fastapi==0.116.2
fastjsonschema==2.21.1
filelock==3.19.1
fire==0.7.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.3
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
llama_stack==0.2.22
llama_stack_client==0.2.22
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.3
ml_dtypes==0.5.3
MouseInfo==0.1.3
multidict==6.6.4
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.1.3
openai==1.107.3
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyaml==25.7.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
pyjokes==0.8.3
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==8.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
scipy==1.16.1
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
Slowloris==0.2.6
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
stack-data==0.6.3
starlette==0.48.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
terminado==0.18.1
tiktoken==0.11.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0


torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
inference-sdk>=0.15.0

